pyproject.toml
[tool.poetry]
name = "redhat_opentelemetry_contrib"
version = "0.0.0"
description = "OpenTelementry Python extensions written at Red Hat"
authors = ["Barak Korren <bkorren@redhat.com>"]
license = "Apache-2.0"

[tool.poetry-dynamic-versioning]
enable = true
metadata = false

[tool.poetry.dependencies]
python = "^3.9.0"
opentelemetry-api = "^1.14.0"
opentelemetry-sdk = "^1.14.0"
opentelemetry-distro = "^0.35b0"
setuptools = "^65.6.3"

[tool.poetry.dev-dependencies]
pylint = "^2.15.6"
black = "^22.10.0"
pypi = "^2.1"
isort = "^5.10.1"
mypy = "^0.991"
pytest = "^7.2.0"
pylint-pytest = "^1.1.2"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.extras]
instruments = []

[tool.poetry.plugins.opentelemetry_traces_exporter]
file = "redhat_opentelemetry_contrib.file_exporters:FileSpanExporter"

[tool.poetry.plugins.opentelemetry_metrics_exporter]
file = "redhat_opentelemetry_contrib.file_exporters:FileMetricExporter"

[tool.poetry.plugins.opentelemetry_instrumentor]
__main__ = "redhat_opentelemetry_contrib.script_instrumentor:ScriptInstrumentor"
